Change next HF version from 3.0 to 4.0.
After external changes, such that v4.0 is the next hydrofabric version
that will be officially released with full support, changing references
here that said 3.0 to 4.0, etc.
