+++ include/core/HY_Features_Ids.hpp
@@ -14,7 +14,7 @@ namespace hy_features {
       static const std::string terminal = "tnx";
       static const std::string catchment = "cat";
       static const std::string flowpath = "wb";    // v2.2 flowpath (waterbody) prefix
-      static const std::string flowpath_v3 = "fp"; // v3.0 flowpath prefix
+      static const std::string flowpath_v4 = "fp"; // v4.0 flowpath prefix
       static const std::string inland = "inx";
 
       /**
@@ -34,7 +34,7 @@ namespace hy_features {
       inline const bool isCatchment(const std::string& type){
         //This is a next-gen specific identifier, hiding it here for now
         static const std::string aggregate = "agg";
-        return (type == catchment || type == aggregate || type == flowpath || type == flowpath_v3);
+        return (type == catchment || type == aggregate || type == flowpath || type == flowpath_v4);
       }
 
       /**
@@ -43,7 +43,7 @@ namespace hy_features {
        * @param type string to test
       */
       inline const bool isFlowpath(const std::string& type){
-        return (type == flowpath || type == flowpath_v3);
+        return (type == flowpath || type == flowpath_v4);
       }
     }
 }
