+++ test/data/geopackage/README.md
@@ -36,11 +36,11 @@ cat-3 (wb-3) ─┘
 
 ---
 
-## example_v3_0_extra_col.gpkg
+## example_v4_0_extra_col.gpkg
 
-Minimal synthetic v3.0 GeoPackage with auxiliary tables carrying extra/unusual
+Minimal synthetic v4.0 GeoPackage with auxiliary tables carrying extra/unusual
 columns. Used by the extra-column tolerance regression tests.
-Generated by `make_v3_0_extra_col_fixture.py`.
+Generated by `make_v4_0_extra_col_fixture.py`.
 
 ### Topology
 
@@ -64,7 +64,7 @@ to V2_2 silently.
 
 ## Modifying fixtures
 
-The five v2.2 / v3.0 `.gpkg` files in this directory are committed
+The five v2.2 / v4.0 `.gpkg` files in this directory are committed
 binaries produced by the matching `make_*_fixture.py` scripts. Tests
 read the committed binaries directly; the scripts are **maintainer-only
 tools** for changing what those binaries contain.
@@ -85,10 +85,10 @@ content change is the bookkeeping noise we'd prefer to avoid.
 
 ```
 python3 make_v2_2_fixture.py            # → example_v2_2.gpkg
-python3 make_v3_0_fixture.py            # → example_v3_0.gpkg
-python3 make_v3_0_minimal_fixture.py    # → example_v3_0_minimal.gpkg
-python3 make_v3_0_dangling_fixture.py   # → example_v3_0_dangling.gpkg
-python3 make_v3_0_extra_col_fixture.py  # → example_v3_0_extra_col.gpkg
+python3 make_v4_0_fixture.py            # → example_v4_0.gpkg
+python3 make_v4_0_minimal_fixture.py    # → example_v4_0_minimal.gpkg
+python3 make_v4_0_dangling_fixture.py   # → example_v4_0_dangling.gpkg
+python3 make_v4_0_extra_col_fixture.py  # → example_v4_0_extra_col.gpkg
 ```
 
 Each script writes its `.gpkg` next to itself. No project venv
@@ -130,7 +130,7 @@ library wrote the file. They affect only the maintainer regen workflow.
 useless. Use the colocated diff tool:
 
 ```
-python3 diff_fixture.py example_v3_0.gpkg
+python3 diff_fixture.py example_v4_0.gpkg
 ```
 
 This compares your working-tree fixture against the committed (HEAD)
