+++ test/data/geopackage/example_v4_0_dangling.gpkg
