+++ test/data/geopackage/example_v4_0_minimal.gpkg
